Chnaged how interrupts are handled

+++ devicesd/include/configuration.h
@@ -5,7 +5,7 @@
 #include <stdint.h>
 
 // Returns CPU int vector or 0 on error
-uint8_t get_ioapic_int(uint32_t intno, uint64_t dest_pid, uint32_t chan);
+uint8_t get_ioapic_int(uint32_t intno, uint64_t dest_pid, uint64_t chan);
 
 // Returns true if was susccessful
 uint8_t configure_interrupts_for(IPC_Reg_Int* desc);
